refactor: split the settings form in two, and drop what nothing reads
The General tab held thirty-eight of the sixty-five controls under one
heading - "Global Options" - that covered twenty-two of them and was
wrong about most: only eight are things a poster can override. Poster
Display Speed sat between the theatre-name block and the header wording,
the Coming Soon text was a hundred and fifty lines from the plate that
renders it, and the header's font and colours were on a different tab
from the header - a seam the documentation had to point at out loud.

Display is now its own screen: Slideshow, Header & name, Poster details.
Appearance sits with the thing it applies to, so the Theme tab dissolves
into it. Settings keeps the three things you set up once - Poster
Sources, Screen power, Account. It is the split docs/ already uses.

Moved to where the rest of their kind live: "Validate movie titles when
syncing" to Poster Sources, where every other sync setting is and where
PublicSettingResource already classified it; HDMI CEC and the on/off
times to a Screen power tab, being the only settings that touch
hardware.

Controls that gate other controls now come first and hide what they
control - "Show speaker config" sat below the two settings it turns on,
and the power times showed whether or not the schedule did.

Dropped five columns nothing reads. show_header_border is the one worth
naming: the plate styles replaced it, the migration before this one
converted every display that had it on to the plaque style, and the
display has read header_style ever since - so the checkbox on the Theme
tab has been inert since that release. The others are a display limit
that was never wired up, two logo flags from when each format had a wide
and a tall variant, and a TMDB v4 key with no field and no caller.

Also here, both found while auditing rather than looked for: MainNav
declared beforeDestroyed, which is not a Vue 3 hook - nor a Vue 2 one,
where it was beforeDestroy - so its resize listener was never removed.
And routes/web.php was ten copies of the same closure, which is why a
new screen works until the first hard refresh; it is a list now.

The open tab lives in the URL rather than in a class the click handler
put on a div, so it survives a reload and can be linked to. That makes a
tab switch a route change, which the unsaved-changes guard would have
caught - it lets a move within the same screen through.
