+++ app/Http/Requests/SettingsRequest.php
@@ -33,15 +33,12 @@ class SettingsRequest extends FormRequest
             'plex_ip_address' => 'required_if:plex_service,true|nullable|string',
             'plex_token' => 'required_if:plex_service,true|nullable|string',
             'poster_display_speed' => 'required|integer|min:2000',
-            'poster_display_limit' => 'nullable',
             'coming_soon_text' => 'nullable|string|max:30',
             'now_playing_text' => 'nullable|string|max:30',
             'show_mpaa_rating' => 'required|boolean',
             'show_audience_rating' => 'required|boolean',
             'show_processing_logos' => 'required|boolean',
-            'show_dolby_atmos_horizontal' => 'required|boolean',
             'show_dolby_atmos_vertical' => 'required|boolean',
-            'show_dolby_vision_horizontal' => 'required|boolean',
             'show_dolby_vision_vertical' => 'required|boolean',
             'show_dts' => 'required|boolean',
             'show_dolby_51' => 'required|boolean',
@@ -74,10 +71,8 @@ class SettingsRequest extends FormRequest
             'kodi_username' => 'required_if:kodi_service,true|nullable|string',
             'kodi_password' => 'required_if:kodi_service,true|nullable|string',
             'tmdb_api_key_v3' => 'nullable|string',
-            'tmdb_api_key_v4' => 'nullable|string',
             'header_bg_color' => 'nullable|string',
             'header_text_color' => 'nullable|string',
-            'show_header_border' => 'required|boolean',
             'header_border_color' => 'nullable|string',
             'header_font' => 'required|string',
             'header_font_size' => 'required|string',
@@ -103,8 +98,6 @@ class SettingsRequest extends FormRequest
             'show_mpaa_rating' => $this->boolean('show_mpaa_rating'),
             'show_audience_rating' => $this->boolean('show_audience_rating'),
             'show_processing_logos' => $this->boolean('show_processing_logos'),
-            'show_dolby_vision_horizontal' => $this->boolean('show_dolby_vision_horizontal'),
-            'show_dolby_atmos_horizontal' => $this->boolean('show_dolby_atmos_horizontal'),
             'show_dolby_atmos_vertical' => $this->boolean('show_dolby_atmos_vertical'),
             'show_dolby_vision_vertical' => $this->boolean('show_dolby_vision_vertical'),
             'show_dts' => $this->boolean('show_dts'),
@@ -132,7 +125,6 @@ class SettingsRequest extends FormRequest
             'poster_fill_scrim' => $this->input('poster_fill_scrim') ?: 'standard',
             'transition_type' => $this->input('transition_type') ?: 'fade',
             'jellyfin_service' => $this->boolean('jellyfin_service'),
-            'show_header_border' => $this->boolean('show_header_border'),
             'validate_movie_titles' => $this->boolean('validate_movie_titles'),
             'remove_black_bars' => $this->boolean('remove_black_bars'),
             'plex_sync_movies' => $this->boolean('plex_sync_movies'),
