+++ tests/Feature/SettingsApiTest.php
@@ -176,12 +176,12 @@ class SettingsApiTest extends TestCase
         $booleans = [
             'plex_service', 'plex_sync_movies', 'plex_sync_tv', 'plex_show_movie_now_playing',
             'plex_show_tv_now_playing', 'random_order', 'show_mpaa_rating', 'show_audience_rating',
-            'show_processing_logos', 'show_dolby_atmos_horizontal', 'show_dolby_atmos_vertical',
-            'show_dolby_vision_horizontal', 'show_dolby_vision_vertical', 'show_dts', 'show_dolby_51',
+            'show_processing_logos', 'show_dolby_atmos_vertical',
+            'show_dolby_vision_vertical', 'show_dts', 'show_dolby_51',
             'show_imax', 'show_auro_3d', 'use_cec_power', 'show_runtime', 'play_theme_music',
             'use_global_prologos', 'use_global_prologos_if_no_poster_prologos', 'poster_fill_screen',
             'show_header_text', 'show_theater_name',
-            'jellyfin_service', 'kodi_service', 'show_header_border', 'validate_movie_titles',
+            'jellyfin_service', 'kodi_service', 'validate_movie_titles',
             'remove_black_bars', 'show_speaker_config',
         ];
